song-service/requirements.txt
yt-dlp>=2023.11.16
syncedlyrics>=0.5.0
pymongo>=4.6.0
python-dotenv>=1.0.0
requests>=2.31.0
spotipy>=2.23.0
flask>=3.0.0
flask-cors>=4.0.0
ytmusicapi>=1.7.0
